Registration for Mentors streamlined
Complete overhaul of how Mentors register. They are now firmly Mentors,
and not a form of User. Adjusted the various models to be compatible
with existing code as well, since there was a lot of discrepancy between
some pages and what the models actually said. Last EJS page is also
removed.